Pull paperdoll-queue out as a separate module
