+++ queue/pom.xml
@@ -9,6 +9,10 @@
 	<artifactId>paperdoll-queue</artifactId>
 	<packaging>jar</packaging>
 	<dependencies>
+        <dependency>
+            <groupId>org.scalaz</groupId>
+            <artifactId>scalaz-core_2.11</artifactId>
+        </dependency>
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
