+++ queue/src/main/scala/paperdoll/queue/Pair.scala
@@ -1,4 +1,4 @@
-package paperdoll.core.queue
+package paperdoll.queue
 
 import scalaz.Forall
